LayoutTests/platform/glib/imported/w3c/web-platform-tests/accname/name/comp_host_language_label-expected.txt

AccName: Host Language Label Tests

Tests the #comp_host_language_label portions of the AccName Name Computation algorithm.

HTML input with value, alt, etc.


HTML input label/for

checkbox label
checkbox label with non-empty value
checkbox label checked
checkbox label checked with non-empty value
color label
color label with non-empty value
date label
date label with non-empty value
datetime-local label
datetime-local label with non-empty value
email label
email label with non-empty value
month label
month label with non-empty value
number label
number label with non-empty value
password label
password label with non-empty value
radio label
radio label with non-empty value
range label
range label with non-empty value
search label
search label with non-empty value
tel label
tel label with non-empty value
textfield label
textfield label with non-empty value
time label
time label with non-empty value
url label
url label with non-empty value
week label
week label with non-empty value
HTML input label encapsulation

checkbox label
checkbox label with non-empty value
checkbox label checked
checkbox label checked with non-empty value
color label
color label with non-empty value
date label
date label with non-empty value
datetime-local label
datetime-local label with non-empty value
email label
email label with non-empty value
month label
month label with non-empty value
number label
number label with non-empty value
password label
password label with non-empty value
radio label
radio label with non-empty value
range label
range label with non-empty value
search label
search label with non-empty value
tel label
tel label with non-empty value
textfield label
textfield label with non-empty value
time label
time label with non-empty value
url label
url label with non-empty value
week label
week label with non-empty value
HTML select

select label
select label
HTML img/picture


HTML fieldset/legend

fieldset legend label

HTML table/caption

table caption label
a	b	c
1	2	3
HTML input with multiple label/for

textfield label 1 textfield label 2

PASS html: input[type=button]
PASS html: input[type=image]
PASS html: input[type=reset]
PASS html: input[type=submit]
PASS html: label[for] input[type=checkbox]
PASS html: label[for] input[type=checkbox][value='test']
PASS html: label[for] input[type=checkbox][checked]
PASS html: label[for] input[type=checkbox][checked][value='test']
PASS html: label[for] input[type=color]
PASS html: label[for] input[type=color][value='#999999']
PASS html: label[for] input[type=date]
PASS html: label[for] input[type=date][value='2025-01-01']
PASS html: label[for] input[type=datetime-local]
PASS html: label[for] input[type=datetime-local][value='2025-01-01T00:01']
PASS html: label[for] input[type=email]
PASS html: label[for] input[type=email][value='test@test.com']
PASS html: label[for] input[type=month]
PASS html: label[for] input[type=month][value='2025-01']
PASS html: label[for] input[type=number]
PASS html: label[for] input[type=number][value=2025]
PASS html: label[for] input[type=password]
PASS html: label[for] input[type=password][value='test']
PASS html: label[for] input[type=radio]
PASS html: label[for] input[type=radio][value='test']
PASS html: label[for] input[type=range]
PASS html: label[for] input[type=range][min=0][max=10][value=5]
PASS html: label[for] input[type=search]
PASS html: label[for] input[type=search][value='test']
PASS html: label[for] input[type=tel]
PASS html: label[for] input[type=tel][value='123-45-678']
PASS html: label[for] input[type=text]
PASS html: label[for] input[type=text][value='test']
PASS html: label[for] input[type=time]
PASS html: label[for] input[type=time][value='00:01']
PASS html: label[for] input[type=url]
PASS html: label[for] input[type=url][value='https://www.w3.org']
PASS html: label[for] input[type=week]
PASS html: label[for] input[type=week][value='2025-W01']
PASS html: label input[type=checkbox] encapsulation
PASS html: label input[type=checkbox][value='test'] encapsulation
PASS html: label input[type=checkbox][checked] encapsulation
PASS html: label input[type=checkbox][value='test'][checked] encapsulation
PASS html: label input[type=color] encapsulation
PASS html: label input[type=color][value='#999999'] encapsulation
PASS html: label input[type=date] encapsulation
FAIL html: label input[type=date][value='2025-01-01'] encapsulation assert_equals: <input type="date" data-expectedlabel="date label with non-empty value" data-testname="html: label input[type=date][value='2025-01-01'] encapsulation" class="ex" value="2025-01-01"> expected "date label with non-empty value" but got "01/01/2025 date label with non-empty value"
PASS html: label input[type=datetime-local] encapsulation
FAIL html: label input[type=datetime-local][value='2025-01-01T00:01'] encapsulation assert_equals: <input type="datetime-local" data-expectedlabel="datetime-local label with non-empty value" data-testname="html: label input[type=datetime-local][value='2025-01-01T00:01'] encapsulation" class="ex" value="2025-01-01T00:01"> expected "datetime-local label with non-empty value" but got "01/01/2025, 12:01:00 AM datetime-local label with non-empty value"
PASS html: label input[type=email] encapsulation
FAIL html: label input[type=email][value='test@test.com'] encapsulation assert_equals: <input type="email" data-expectedlabel="email label with non-empty value" data-testname="html: label input[type=email][value='test@test.com'] encapsulation" class="ex" value="test@test.com"> expected "email label with non-empty value" but got "test@test.com email label with non-empty value"
PASS html: label input[type=month] encapsulation
FAIL html: label input[type=month][value='2025-01'] encapsulation assert_equals: <input type="month" data-expectedlabel="month label with non-empty value" data-testname="html: label input[type=month][value='2025-01'] encapsulation" class="ex" value="2025-01"> expected "month label with non-empty value" but got "January 2025 month label with non-empty value"
PASS html: label input[type=number] encapsulation
FAIL html: label input[type=number][value=1] encapsulation assert_equals: <input type="number" data-expectedlabel="number label with non-empty value" data-testname="html: label input[type=number][value=1] encapsulation" class="ex" value="1"> expected "number label with non-empty value" but got "1 number label with non-empty value"
PASS html: label input[type=password] encapsulation
FAIL html: label input[type=password][value='test'] encapsulation assert_equals: <input type="password" data-expectedlabel="password label with non-empty value" data-testname="html: label input[type=password][value='test'] encapsulation" class="ex" value="test"> expected "password label with non-empty value" but got "•••• password label with non-empty value"
PASS html: label input[type=radio] encapsulation
PASS html: label input[type=radio][value='test'] encapsulation
FAIL html: label input[type=range] encapsulation assert_equals: <input type="range" data-expectedlabel="range label" data-testname="html: label input[type=range] encapsulation" class="ex"> expected "range label" but got "50 range label"
FAIL html: label input[type=range][value='5'][min='0'][max='10'] encapsulation assert_equals: <input type="range" data-expectedlabel="range label with non-empty value" data-testname="html: label input[type=range][value='5'][min='0'][max='10'] encapsulation" class="ex" min="0" max="10" value="5"> expected "range label with non-empty value" but got "5 range label with non-empty value"
FAIL html: label input[type=search] encapsulation assert_equals: <input type="search" data-expectedlabel="search label" data-testname="html: label input[type=search] encapsulation" class="ex"> expected "search label" but got "cancel search label"
FAIL html: label input[type=search][value='test'] encapsulation assert_equals: <input type="search" data-expectedlabel="search label with non-empty value" data-testname="html: label input[type=search][value='test'] encapsulation" class="ex" value="test"> expected "search label with non-empty value" but got "test cancel search label with non-empty value"
PASS html: label input[type=tel] encapsulation
FAIL html: label input[type=tel][value='123-45-678'] encapsulation assert_equals: <input type="tel" data-expectedlabel="tel label with non-empty value" data-testname="html: label input[type=tel][value='123-45-678'] encapsulation" class="ex" value="123-45-678"> expected "tel label with non-empty value" but got "123-45-678 tel label with non-empty value"
PASS html: label[for] input[type=text] encapsulation
FAIL html: label[for] input[type=text][value='test'] encapsulation assert_equals: <input type="text" data-expectedlabel="textfield label with non-empty value" data-testname="html: label[for] input[type=text][value='test'] encapsulation" class="ex" value="test"> expected "textfield label with non-empty value" but got "test textfield label with non-empty value"
PASS html: label input[type=time] encapsulation
FAIL html: label input[type=time][value='00:01'] encapsulation assert_equals: <input type="time" data-expectedlabel="time label with non-empty value" data-testname="html: label input[type=time][value='00:01'] encapsulation" class="ex" value="00:01"> expected "time label with non-empty value" but got "12:01:00 AM time label with non-empty value"
PASS html: label input[type=url] encapsulation
FAIL html: label input[type=url][value='https://www.w3.org'] encapsulation assert_equals: <input type="url" data-expectedlabel="url label with non-empty value" data-testname="html: label input[type=url][value='https://www.w3.org'] encapsulation" class="ex" value="https://www.w3.org"> expected "url label with non-empty value" but got "https://www.w3.org url label with non-empty value"
PASS html: label input[type=week] encapsulation
FAIL html: label input[type=week][value='2025-W01'] encapsulation assert_equals: <input type="week" data-expectedlabel="week label with non-empty value" data-testname="html: label input[type=week][value='2025-W01'] encapsulation" class="ex" value="2025-W01"> expected "week label with non-empty value" but got "2025-W01 week label with non-empty value"
PASS html: select for/id
FAIL html: select encapsulation assert_equals: <select data-expectedlabel="select label" data-testname="html: select encapsulation" class="ex">
    <option>foo</option>
  </select> expected "select label" but got "select label foo"
PASS html: img[alt] (non-empty)
PASS html: picture > img[alt] (non-empty)
PASS img[alt] with single space character as name
PASS img[alt] with single tab as name
PASS img[alt] with new line as name
PASS img[alt] with single space character and tab as name
PASS img[alt] with form feed as name
PASS img[alt] with interior single tab in the name
PASS img[alt] with interior new line in the name
PASS img[alt] with interior single space and tab in the name
PASS img[alt] with interior form feed in the name
PASS html: fieldset > legend
PASS html: table > caption
PASS html: multiple label elements[for]
